requirements.txt
# Core dependencies
feedparser>=6.0.10
beautifulsoup4>=4.12.0
lxml>=4.9.0
pdfplumber>=0.10.0
openai>=1.0.0
sqlalchemy>=2.0.0
alembic>=1.12.0
psycopg2-binary>=2.9.9
python-dotenv>=1.0.0
requests>=2.31.0

# Dashboard framework
flask>=3.0.0
flask-cors>=4.0.0

# Utilities
python-dateutil>=2.8.2
pytz>=2023.3

# Development dependencies
pytest>=7.4.0
pytest-cov>=4.1.0
black>=23.12.0
flake8>=6.1.0

